opt/stacks/adversarial_playground_ai_target/adversarial-playground/requirements.txt
numpy 
matplotlib
pandas
Pillow==9.2.0 
tensorflow==2.15.*
ipykernel
flask
scikit-learn